bot/requirements.txt
aiogram>=3.0.0
python-dotenv>=0.19.0
aiohttp>=3.8.0